requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.0
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pycparser==2.22
pydotplus==2.0.2
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python-decouple==3.8
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
